Moved interfaces to the doc folder
