+++ packages/vsce/src/doc/commands/ICommandParams.ts
