+++ packages/vsce/src/doc/commands/IFocusRegion.ts
@@ -9,11 +9,12 @@
  *
  */
 
-import { IProfile } from "@zowe/imperative";
-import { CICSSession } from "./CICSSession";
+import { IProfileLoaded } from "@zowe/imperative";
+import { CICSSession } from "../../resources/CICSSession";
 
-export interface ISessionHandler {
-  getSession(profile: IProfile): CICSSession | undefined;
-  removeSession(profileName: string): void;
-  clearSessions(): void;
+export interface IFocusRegion {
+  profile: IProfileLoaded;
+  cicsPlex?: string;
+  session: CICSSession;
+  focusSelectedRegion: string;
 }
